+++ vrepsim/objects.py
@@ -580,6 +580,28 @@ class VisionSensor(SceneObject):
                                                   # decimal point)
         return clip_plane
 
+    def set_far_clip_plane(self, clip_plane):
+        """Set far clipping plane."""
+        if self._handle < 0:
+            if self._handle == MISSING_HANDLE:
+                raise RuntimeError(
+                    "Could not set far clipping plane of {}: missing name or "
+                    "handle.".format(self._name))
+            if self._handle == REMOVED_OBJ_HANDLE:
+                raise RuntimeError("Could not set far clipping plane of {}: "
+                                   "object removed.".format(self._name))
+        client_id = self.client_id
+        if client_id is None:
+            raise ConnectionError(
+                "Could not set far clipping plane of {}: not connected to "
+                "V-REP remote API server.".format(self._name))
+        res = vrep.simxSetObjectFloatParameter(
+            client_id, self._handle, vrep.sim_visionfloatparam_far_clipping,
+            clip_plane, vrep.simx_opmode_blocking)
+        if res != vrep.simx_return_ok:
+            raise ServerError("Could not set far clipping plane of {}."
+                              "".format(self._name))
+
     def get_image(self, grayscale=False):
         """Retrieve image."""
         # Retrieve image from the vision sensor simulated in V-REP