+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Changelog
   sensor simulated in V-REP.
 - Added retrieving vision sensor far clipping plane via interface to vision
   sensor simulated in V-REP.
+- Added setting vision sensor far clipping plane via interface to vision sensor
+  simulated in V-REP.
 
 ### Changed